Add all java files to "aisu" package
