+++ src/main/java/aisu/Aisu.java
@@ -1,3 +1,4 @@
+package aisu;
 import java.util.List;
 
 /**
