+++ src/main/java/aisu/Parser.java
@@ -1,3 +1,4 @@
+package aisu;
 /**
  * The Parser class deals with making sense of the user command.
  */
