+++ src/main/java/aisu/Storage.java
@@ -1,3 +1,5 @@
+package aisu;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
