+++ src/main/java/aisu/Task.java
@@ -1,3 +1,4 @@
+package aisu;
 /**
  * This is the Task class.
  * A Task contains a description and a status of being done or not.
