+++ src/main/java/aisu/Todo.java
@@ -1,3 +1,4 @@
+package aisu;
 public class Todo extends Task {
     public Todo(String description) throws AisuException {
         super(description);
