+++ src/main/java/aisu/Ui.java
@@ -1,3 +1,4 @@
+package aisu;
 import java.io.PrintStream;
 import java.util.Scanner;
